Adjusted some classes, specifically added list of books and changed the tipe of ISBN to Long to avoid errors

+++ src/mpp/aed/library/Library.java
@@ -91,7 +91,7 @@ public class Library implements Serializable {
         return false;
     }
     
-    public Book getBookByISBN(int ISBN) {
+    public Book getBookByISBN(long ISBN) {
         for( Book book : books ) {
             if( book.getISBN() == ISBN ) {
                 return book;
